% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anecdotes2.R
\docType{package}
\name{anecdotes2}
\alias{anecdotes2}
\alias{anecdotes2-package}
\title{Package Title}
\description{
A description of the data package
}
\details{
anecdotes2
A data package for anecdotes2.

Use \code{data(package='anecdotes2')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{data}
}
