% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attention_check.R
\name{attention_check}
\alias{attention_check}
\title{Test for attention check

If more than one fail, remove data, and reject submission.

According to Prolific
"If a participant fails more than one fair attention check, they can justifiably be rejected."
https://researcher-help.prolific.co/hc/en-gb/articles/360009223553-

Output submissions for rejection.

Warn if any fail.

If project test fails, remove submission even if it is the only one, but do not reject.}
\usage{
attention_check(data, test_name, test_answer)
}
\value{

}
\description{
Test for attention check

If more than one fail, remove data, and reject submission.

According to Prolific
"If a participant fails more than one fair attention check, they can justifiably be rejected."
https://researcher-help.prolific.co/hc/en-gb/articles/360009223553-

Output submissions for rejection.

Warn if any fail.

If project test fails, remove submission even if it is the only one, but do not reject.
}
\author{
Shir Dekel
}
