% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_when_target.R
\name{case_when_target}
\alias{case_when_target}
\title{Conditional target processing

Include all components for anecdotes and combined conditions, but no
statistics for anecdote only condition.}
\usage{
case_when_target(anecdote_within, anecdote_between, data)
}
\value{

}
\description{
Conditional target processing

Include all components for anecdotes and combined conditions, but no
statistics for anecdote only condition.
}
\author{
Shir Dekel
}
