% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_server_aggregation_exp2.R
\name{download_data_server_aggregation_exp2}
\alias{download_data_server_aggregation_exp2}
\title{Download server data}
\usage{
download_data_server_aggregation_exp2()
}
\value{

}
\description{
Download server data
}
\author{
Shir Dekel
}
