% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data_raw_jspsych.R
\name{filter_data_raw_jspsych}
\alias{filter_data_raw_jspsych}
\title{Filter raw data

Take the full csv from the server and split it up according to the relevant
experiment.

Previously, downloading new data meant that every experiment reran data
cleaning}
\usage{
filter_data_raw_jspsych(data_raw, thesis_project, experiment_number)
}
\value{

}
\description{
Filter raw data

Take the full csv from the server and split it up according to the relevant
experiment.

Previously, downloading new data meant that every experiment reran data
cleaning
}
\author{
Shir Dekel
}
