% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_within_subjects_components.R
\name{fix_within_subjects_components}
\alias{fix_within_subjects_components}
\title{Fix within-subjects components
Make sure positive valence conditions have a low statistics and negative
valence conditions have high statistics.

Make sure that structure and integration match the relevant anecdote
condition.}
\usage{
fix_within_subjects_components(within_subjects_unnested)
}
\value{

}
\description{
Fix within-subjects components
Make sure positive valence conditions have a low statistics and negative
valence conditions have high statistics.

Make sure that structure and integration match the relevant anecdote
condition.
}
\author{
Shir Dekel
}
