% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_results_alignment_6.R
\name{get_all_results_alignment_6}
\alias{get_all_results_alignment_6}
\title{All results}
\usage{
get_all_results_alignment_6(data, dv)
}
\value{

}
\description{
All results
}
\author{
Shir Dekel
}
