% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analysis.R
\name{get_analysis}
\alias{get_analysis}
\title{Get anecdote analysis}
\usage{
get_analysis(
  business_name,
  success,
  reason_location,
  location,
  structure,
  reason_structure,
  integration,
  reason_integration,
  reason_cutoff,
  value_string,
  reason_value_string,
  value_numeric,
  cutoff,
  unit
)
}
\value{

}
\description{
Get anecdote analysis
}
\author{
Shir Dekel
}
