% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_anecdote.R
\name{get_anecdote}
\alias{get_anecdote}
\title{Get anecdote html
that'll end up in "Case study"}
\usage{
get_anecdote(data)
}
\value{

}
\description{
Get anecdote html
that'll end up in "Case study"
}
\author{
Shir Dekel
}
