% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_anecdote_raw.R
\name{get_anecdote_raw}
\alias{get_anecdote_raw}
\title{Get raw anecdote description

Need to remove whitespace so that when concatenated by \code{insert_javascript}
it stays on the same line.}
\usage{
get_anecdote_raw(
  business_name_anecdote,
  location_anecdote,
  integration_anecdote,
  structure_anecdote,
  type_anecdote,
  predicted_features_anecdote
)
}
\value{

}
\description{
Get raw anecdote description

Need to remove whitespace so that when concatenated by \code{insert_javascript}
it stays on the same line.
}
\author{
Shir Dekel
}
