% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_anecdotes_only_similarity.R
\name{get_anecdotes_only_similarity}
\alias{get_anecdotes_only_similarity}
\title{Get similarity analysis in the anecdotes only condition

For Anecdotes 2

Similarity check
Negative valence
Anecdotes only: low similarity > high similarity
Positive valence
Anecdotes only: high similarity > low similarity}
\usage{
get_anecdotes_only_similarity(model)
}
\value{

}
\description{
Get similarity analysis in the anecdotes only condition

For Anecdotes 2

Similarity check
Negative valence
Anecdotes only: low similarity > high similarity
Positive valence
Anecdotes only: high similarity > low similarity
}
\author{
Shir Dekel
}
