% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_business_name_anecdotes_2.R
\name{get_business_name_anecdotes_2}
\alias{get_business_name_anecdotes_2}
\title{Busines names anecdotes 2
Names generated from namelix.com}
\usage{
get_business_name_anecdotes_2()
}
\value{

}
\description{
Busines names anecdotes 2
Names generated from namelix.com
}
\author{
Shir Dekel
}
