% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_combined.R
\name{get_combined}
\alias{get_combined}
\title{Get combined anecdote condition analysis
Anecdotal bias moderated by similarity
Negative valence
Statistics only > combined high similarity
Combined: Low similarity > high similarity
Statistics only = combined low similarity
Positive valence
Statistics only > combined low similarity
Combined: High similarity > low similarity
Statistics only > combined high similarity}
\usage{
get_combined(data, valence)
}
\value{

}
\description{
Get combined anecdote condition analysis
Anecdotal bias moderated by similarity
Negative valence
Statistics only > combined high similarity
Combined: Low similarity > high similarity
Statistics only = combined low similarity
Positive valence
Statistics only > combined low similarity
Combined: High similarity > low similarity
Statistics only > combined high similarity
}
\author{
Shir Dekel
}
