% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_condition_allocation.R
\name{get_condition_allocation}
\alias{get_condition_allocation}
\title{Parse condition allocation JS code for anecdotes 2}
\usage{
get_condition_allocation()
}
\value{

}
\description{
Parse condition allocation JS code for anecdotes 2
}
\author{
Shir Dekel
}
