% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conditions.R
\name{get_conditions}
\alias{get_conditions}
\title{Get conditions dataframe for anecdotes 2}
\usage{
get_conditions()
}
\value{

}
\description{
Get conditions dataframe for anecdotes 2
}
\author{
Shir Dekel
}
