% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_simulation.R
\name{get_data_simulation}
\alias{get_data_simulation}
\title{Data simulation for anecdotes 2}
\usage{
get_data_simulation()
}
\value{

}
\description{
Data simulation for anecdotes 2
}
\author{
Shir Dekel
}
