% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_simulation_results_anecdotes_2.R
\name{get_data_simulation_results_anecdotes_2}
\alias{get_data_simulation_results_anecdotes_2}
\title{Anecdote 2 simulation results}
\usage{
get_data_simulation_results_anecdotes_2(data)
}
\value{

}
\description{
Anecdote 2 simulation results
}
\author{
Shir Dekel
}
