% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df_anecdotes_2.R
\name{get_df_anecdotes_2}
\alias{get_df_anecdotes_2}
\title{Get anecdotes 2 simulation df}
\usage{
get_df_anecdotes_2(n = 10)
}
\arguments{
\item{n}{must be divisible by 2}
}
\value{

}
\description{
Get anecdotes 2 simulation df
}
\author{
Shir Dekel
}
