% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_display_anecdotes_2.R
\name{get_display_anecdotes_2}
\alias{get_display_anecdotes_2}
\title{Get anecdotes 2 display

Statistics only condition only sees the target project display, whereas
those in the anecdotes conditions (anecdote_only and combined) see both
target and anecdote description.}
\usage{
get_display_anecdotes_2(instructions, anecdote_within, target, anecdote)
}
\value{

}
\description{
Get anecdotes 2 display

Statistics only condition only sees the target project display, whereas
those in the anecdotes conditions (anecdote_only and combined) see both
target and anecdote description.
}
\author{
Shir Dekel
}
