% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_display_setup.R
\name{get_display_setup}
\alias{get_display_setup}
\title{Set up parameters for display processing}
\usage{
get_display_setup()
}
\value{

}
\description{
Set up parameters for display processing
}
\author{
Shir Dekel
}
