% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_name_materials_anecdotes_2.R
\name{get_file_name_materials_anecdotes_2}
\alias{get_file_name_materials_anecdotes_2}
\title{Anecdotes 2 screenshot file names

Get all displays from the combined condition, interstitials, and follow_ups,
with a sample anecdotes only display at the end.

Don't need to include variations because everyone sees the same domains; the
variation is in the associated conditions.}
\usage{
get_file_name_materials_anecdotes_2()
}
\value{

}
\description{
Anecdotes 2 screenshot file names

Get all displays from the combined condition, interstitials, and follow_ups,
with a sample anecdotes only display at the end.

Don't need to include variations because everyone sees the same domains; the
variation is in the associated conditions.
}
\author{
Shir Dekel
}
