% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_follow_up.R
\name{get_follow_up}
\alias{get_follow_up}
\title{Get follow-up trials

For anecdote displays}
\usage{
get_follow_up(anecdote_within, data, valence, similarity)
}
\value{

}
\description{
Get follow-up trials

For anecdote displays
}
\author{
Shir Dekel
}
