% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_follow_up_html.R
\name{get_follow_up_html}
\alias{get_follow_up_html}
\title{Get follow-up question HTML

Identify similarity and relevance ratings}
\usage{
get_follow_up_html(data, button_continue, anecdote_within, valence, similarity)
}
\value{

}
\description{
Get follow-up question HTML

Identify similarity and relevance ratings
}
\author{
Shir Dekel
}
