% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_follow_up_textarea.R
\name{get_follow_up_textarea}
\alias{get_follow_up_textarea}
\title{Get textarea for justify follow-up}
\usage{
get_follow_up_textarea(follow_up_id)
}
\value{

}
\description{
Get textarea for justify follow-up
}
\author{
Shir Dekel
}
