% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_id.R
\name{get_input_id}
\alias{get_input_id}
\title{Get anecdotes 2 input ID}
\usage{
get_input_id(
  feature,
  value,
  anecdote_within,
  similarity,
  valence,
  business_name,
  type,
  npv,
  reliability,
  project_type
)
}
\value{

}
\description{
Get anecdotes 2 input ID
}
\author{
Shir Dekel
}
