% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instructions_anecdotes_2.R
\name{get_instructions_anecdotes_2}
\alias{get_instructions_anecdotes_2}
\title{Instructions anecdotes 2}
\usage{
get_instructions_anecdotes_2()
}
\value{

}
\description{
Instructions anecdotes 2
}
\author{
Shir Dekel
}
