% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instructions_specific.R
\name{get_instructions_specific}
\alias{get_instructions_specific}
\title{Get condition-specific instructions}
\usage{
get_instructions_specific(anecdote_within, anecdote_between, success)
}
\value{

}
\description{
Get condition-specific instructions
}
\author{
Shir Dekel
}
