% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interstitial_anecdotes.R
\name{get_interstitial_anecdotes}
\alias{get_interstitial_anecdotes}
\title{Get interstitial 1 for anecdotes 2}
\usage{
get_interstitial_anecdotes(interstitial_number)
}
\value{

}
\description{
Get interstitial 1 for anecdotes 2
}
\author{
Shir Dekel
}
