% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_location_anecdotes_2.R
\name{get_location_anecdotes_2}
\alias{get_location_anecdotes_2}
\title{Anecdotes 2 location}
\usage{
get_location_anecdotes_2()
}
\value{

}
\description{
Anecdotes 2 location
}
\author{
Shir Dekel
}
