% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_main.R
\name{get_main}
\alias{get_main}
\title{Get main experiment trials for anecdotes 2

Should be a list of one timeline. That timeline should be a list of n, with
n being the number of trials. If a trial is a list (because it is a
conditional timeline or has variables), then it should be a timeline list.

Usually you're using lists of trials, so running \code{map(flatten)} usually does
the trick.}
\usage{
get_main(gambles, randomize_order = TRUE)
}
\arguments{
\item{gambles}{}

\item{randomize_order}{}
}
\value{

}
\description{
Get main experiment trials for anecdotes 2

Should be a list of one timeline. That timeline should be a list of n, with
n being the number of trials. If a trial is a list (because it is a
conditional timeline or has variables), then it should be a timeline list.

Usually you're using lists of trials, so running \code{map(flatten)} usually does
the trick.
}
\author{
Shir Dekel
}
