% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_anecdotes_2.R
\name{get_model_anecdotes_2}
\alias{get_model_anecdotes_2}
\title{Get anecdotes 2 model}
\usage{
get_model_anecdotes_2(data_clean)
}
\value{

}
\description{
Get anecdotes 2 model
}
\author{
Shir Dekel
}
