% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multiplier_anecdote.R
\name{get_multiplier_anecdote}
\alias{get_multiplier_anecdote}
\title{Get multipliers for anecdotes 2}
\usage{
get_multiplier_anecdote()
}
\value{

}
\description{
Get multipliers for anecdotes 2
}
\author{
Shir Dekel
}
