% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nesting_setup.R
\name{get_nesting_setup}
\alias{get_nesting_setup}
\title{Get nesting set up

Nest in order so that elements can be subsequently added for each relevant
layer (project type, feature type, and within subjects conditions).}
\usage{
get_nesting_setup(conditions)
}
\value{

}
\description{
Get nesting set up

Nest in order so that elements can be subsequently added for each relevant
layer (project type, feature type, and within subjects conditions).
}
\author{
Shir Dekel
}
