% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnibus_allocation.R
\name{get_omnibus_allocation}
\alias{get_omnibus_allocation}
\title{Get omnibus analysis for allocation data}
\usage{
get_omnibus_allocation(data = anecdotes2::data)
}
\value{

}
\description{
Get omnibus analysis for allocation data
}
\author{
Shir Dekel
}
