% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnibus_allocation_difference.R
\name{get_omnibus_allocation_difference}
\alias{get_omnibus_allocation_difference}
\title{Get omnibus analysis for allocation difference data}
\usage{
get_omnibus_allocation_difference(data = anecdotes2::data)
}
\value{

}
\description{
Get omnibus analysis for allocation difference data
}
\author{
Shir Dekel
}
