% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters_anecdotes_2.R
\name{get_parameters_anecdotes_2}
\alias{get_parameters_anecdotes_2}
\title{Get anecdotes 2 parameters

Generate experiment condition structure and relevant parameters for each
condition. For subsequent parsing into javascript using jaysire.

Used for display and follow_up columns}
\usage{
get_parameters_anecdotes_2()
}
\value{

}
\description{
Get anecdotes 2 parameters

Generate experiment condition structure and relevant parameters for each
condition. For subsequent parsing into javascript using jaysire.

Used for display and follow_up columns
}
\author{
Shir Dekel
}
