% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_simulation.R
\name{get_plot_simulation}
\alias{get_plot_simulation}
\title{Anecdotes simulation plot}
\usage{
get_plot_simulation(data_clean_anecdotes_2)
}
\value{

}
\description{
Anecdotes simulation plot
}
\author{
Shir Dekel
}
