% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_power_table.R
\name{get_power_table}
\alias{get_power_table}
\title{Get power table

The N we are interested in is the per group sample size

So the input of n_vector is in terms of group sample size, but we then need
to multiply it by 2 (two between-subjects conditions) in order to input it
to sim_design, which divides it by 2 so you can get the correct full sample
size.}
\usage{
get_power_table()
}
\value{

}
\description{
Get power table

The N we are interested in is the per group sample size

So the input of n_vector is in terms of group sample size, but we then need
to multiply it by 2 (two between-subjects conditions) in order to input it
to sim_design, which divides it by 2 so you can get the correct full sample
size.
}
\author{
Shir Dekel
}
