% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predicted_features.R
\name{get_predicted_features}
\alias{get_predicted_features}
\title{Get predicted features

Anecdotes 2}
\usage{
get_predicted_features(value, feature, unit)
}
\value{

}
\description{
Get predicted features

Anecdotes 2
}
\author{
Shir Dekel
}
