% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_type_anecdotes_2.R
\name{get_project_type_anecdotes_2}
\alias{get_project_type_anecdotes_2}
\title{Project type}
\usage{
get_project_type_anecdotes_2()
}
\value{

}
\description{
Project type
}
\author{
Shir Dekel
}
