% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_value_base.R
\name{get_project_value_base}
\alias{get_project_value_base}
\title{Project base value for anecdotes 2}
\usage{
get_project_value_base()
}
\value{

}
\description{
Project base value for anecdotes 2
}
\author{
Shir Dekel
}
