% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projects_anecdotes_2.R
\name{get_projects_anecdotes_2}
\alias{get_projects_anecdotes_2}
\title{Get anecdotes 2 projects

Parse components as javascript with relevant conditional statements to
correspond with the between-subjects conditions.}
\usage{
get_projects_anecdotes_2(
  project_variation,
  anecdote_variation,
  anecdote_between,
  data,
  randomize_order
)
}
\value{

}
\description{
Get anecdotes 2 projects

Parse components as javascript with relevant conditional statements to
correspond with the between-subjects conditions.
}
\author{
Shir Dekel
}
