% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reason_integration.R
\name{get_reason_integration}
\alias{get_reason_integration}
\title{Get integration reason}
\usage{
get_reason_integration()
}
\value{

}
\description{
Get integration reason
}
\author{
Shir Dekel
}
