% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reason_location.R
\name{get_reason_location}
\alias{get_reason_location}
\title{Get location reason

Varies by valence}
\usage{
get_reason_location()
}
\value{

}
\description{
Get location reason

Varies by valence
}
\author{
Shir Dekel
}
