% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reason_structure.R
\name{get_reason_structure}
\alias{get_reason_structure}
\title{Get structure reason

Varies by valence}
\usage{
get_reason_structure()
}
\value{

}
\description{
Get structure reason

Varies by valence
}
\author{
Shir Dekel
}
