% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reason_value_string.R
\name{get_reason_value_string}
\alias{get_reason_value_string}
\title{Reason for string value}
\usage{
get_reason_value_string()
}
\value{

}
\description{
Reason for string value
}
\author{
Shir Dekel
}
