% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reason_value_string_negative.R
\name{get_reason_value_string_negative}
\alias{get_reason_value_string_negative}
\title{Reasons for string value - negative valence}
\usage{
get_reason_value_string_negative()
}
\value{

}
\description{
Reasons for string value - negative valence
}
\author{
Shir Dekel
}
