% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reason_value_string_positive.R
\name{get_reason_value_string_positive}
\alias{get_reason_value_string_positive}
\title{Reasons for string value - positive valence}
\usage{
get_reason_value_string_positive()
}
\value{

}
\description{
Reasons for string value - positive valence
}
\author{
Shir Dekel
}
