% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reliability_anecdotes_2.R
\name{get_reliability_anecdotes_2}
\alias{get_reliability_anecdotes_2}
\title{Reliability anecdotes 2}
\usage{
get_reliability_anecdotes_2()
}
\value{

}
\description{
Reliability anecdotes 2
}
\author{
Shir Dekel
}
