% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_similarity_high_anecdote.R
\name{get_similarity_high_anecdote}
\alias{get_similarity_high_anecdote}
\title{Get anecdote analysis in the high similarity condition

For Anecdotes 2
Effect of statistics
Negative valence
High similarity: combined > anecdote only
Positive valence
High similarity: anecdote only > combined}
\usage{
get_similarity_high_anecdote(model)
}
\value{

}
\description{
Get anecdote analysis in the high similarity condition

For Anecdotes 2
Effect of statistics
Negative valence
High similarity: combined > anecdote only
Positive valence
High similarity: anecdote only > combined
}
\author{
Shir Dekel
}
