% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simulation_summary_anecdotes_2.R
\name{get_simulation_summary_anecdotes_2}
\alias{get_simulation_summary_anecdotes_2}
\title{Anecdotes 2 simulation summary

For power analysis}
\usage{
get_simulation_summary_anecdotes_2(n = 10, nsim)
}
\value{

}
\description{
Anecdotes 2 simulation summary

For power analysis
}
\author{
Shir Dekel
}
