% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_success.R
\name{get_success}
\alias{get_success}
\title{Get success description

Varies by valence condition

1-3 for anecdote.
4-5 for instructions.}
\usage{
get_success()
}
\value{

}
\description{
Get success description

Varies by valence condition

1-3 for anecdote.
4-5 for instructions.
}
\author{
Shir Dekel
}
