% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeline.R
\name{get_timeline}
\alias{get_timeline}
\title{Get anecdotes 2 timeline

Parse each between-subjects row (with each of the five within-subjects
display within the data column) as javascript using jaysire.}
\usage{
get_timeline(parameters_anecdotes_2, randomize_order)
}
\value{

}
\description{
Get anecdotes 2 timeline

Parse each between-subjects row (with each of the five within-subjects
display within the data column) as javascript using jaysire.
}
\author{
Shir Dekel
}
