% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_value_string.R
\name{get_value_string}
\alias{get_value_string}
\title{String comparison value

from Anecdotes 2}
\usage{
get_value_string()
}
\value{

}
\description{
String comparison value

from Anecdotes 2
}
\author{
Shir Dekel
}
