% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_with_datetime.R
\name{import_with_datetime}
\alias{import_with_datetime}
\title{Import with datetime

For Anecdotes 2

File name has the datetime}
\usage{
import_with_datetime(file_path)
}
\value{

}
\description{
Import with datetime

For Anecdotes 2

File name has the datetime
}
\author{
Shir Dekel
}
