% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jspsych_sample.R
\name{jspsych_sample}
\alias{jspsych_sample}
\title{Generate jsPsych sample with replacement}
\usage{
jspsych_sample(condition_level, condition_name)
}
\value{

}
\description{
Generate jsPsych sample with replacement
}
\author{
Shir Dekel
}
