% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latin_list.R
\name{latin_list}
\alias{latin_list}
\title{Generate a latin square design as a list}
\usage{
latin_list(x)
}
\arguments{
\item{x}{A vector or list.}
}
\value{
A list.
}
\description{
Generate a latin square design as a list
}
\examples{
x <- c(1:4)
latin_list(x)
}
