% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_between_subjects.R
\name{mutate_between_subjects}
\alias{mutate_between_subjects}
\title{Mutate between subjects components

Add elements that vary in within-subjects conditions and are identical
across between subjects. This is done so that they can be subsequently
unnested along with ,the within-subjects condition labels.}
\usage{
mutate_between_subjects(varation_mutated)
}
\value{

}
\description{
Mutate between subjects components

Add elements that vary in within-subjects conditions and are identical
across between subjects. This is done so that they can be subsequently
unnested along with ,the within-subjects condition labels.
}
\author{
Shir Dekel
}
