% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_project_variation.R
\name{mutate_project_variation}
\alias{mutate_project_variation}
\title{Mutate project variation components

Mutate business name, type, location, etc. here so that each of the five
sets (of four) of these are associated with a project variation number.}
\usage{
mutate_project_variation(between_subjects_unnested)
}
\value{

}
\description{
Mutate project variation components

Mutate business name, type, location, etc. here so that each of the five
sets (of four) of these are associated with a project variation number.
}
\author{
Shir Dekel
}
