% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_variation.R
\name{mutate_variation}
\alias{mutate_variation}
\title{Mutate variation conditions
Mutate project variation here so that each between subjects condition is
identical and each gets the set of five counterbalanced sequences. Do the
same with anecdot e variation.}
\usage{
mutate_variation(nesting_setup)
}
\value{

}
\description{
Mutate variation conditions
Mutate project variation here so that each between subjects condition is
identical and each gets the set of five counterbalanced sequences. Do the
same with anecdot e variation.
}
\author{
Shir Dekel
}
