% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_for_display.R
\name{nest_for_display}
\alias{nest_for_display}
\title{Nest for subsequent display

Organise table such that each row is a distinct condition, and the data
column consists of the two project as rows, with target and anecdote columns
for subsequent processing.}
\usage{
nest_for_display(feature_type_pivoted)
}
\value{

}
\description{
Nest for subsequent display

Organise table such that each row is a distinct condition, and the data
column consists of the two project as rows, with target and anecdote columns
for subsequent processing.
}
\author{
Shir Dekel
}
