% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_for_timeline.R
\name{nest_for_timeline}
\alias{nest_for_timeline}
\title{Nest for subsequent timeline generation

Nest once to group within-subjects conditions with their relevant displays
and then nest again to group between-subjects conditions with their relevant
within-subjects conditions.}
\usage{
nest_for_timeline(parameters_anecdotes_2)
}
\value{

}
\description{
Nest for subsequent timeline generation

Nest once to group within-subjects conditions with their relevant displays
and then nest again to group between-subjects conditions with their relevant
within-subjects conditions.
}
\author{
Shir Dekel
}
