% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_feature_type.R
\name{pivot_feature_type}
\alias{pivot_feature_type}
\title{Pivot feature type

Makes the feature type rows (anecdote and target) into columns (for the
relevant components) so that conditions (represented by rows) can have
clearly connected anecdote and target components.}
\usage{
pivot_feature_type(unselected_for_pivot)
}
\value{

}
\description{
Pivot feature type

Makes the feature type rows (anecdote and target) into columns (for the
relevant components) so that conditions (represented by rows) can have
clearly connected anecdote and target components.
}
\author{
Shir Dekel
}
