% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_allocation_anecdotes_2.R
\name{plot_allocation_anecdotes_2}
\alias{plot_allocation_anecdotes_2}
\title{Plot anecdotes 2 allocation}
\usage{
plot_allocation_anecdotes_2(data_clean_anecdotes_2, valence)
}
\value{

}
\description{
Plot anecdotes 2 allocation
}
\author{
Shir Dekel
}
