% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_allocation_difference_anecdotes_2.R
\name{plot_allocation_difference_anecdotes_2}
\alias{plot_allocation_difference_anecdotes_2}
\title{Plot allocation difference}
\usage{
plot_allocation_difference_anecdotes_2(omnibus)
}
\value{

}
\description{
Plot allocation difference
}
\author{
Shir Dekel
}
