% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anecdotes2.R
\docType{data}
\name{plot_simulation}
\alias{plot_simulation}
\title{Descriptive data title}
\format{
a \code{list} containing the following fields:
\describe{
\item{negative}{}
\item{positive}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{anecdotes2}
\link{power}
}
