% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_simulation.R
\name{summarise_simulation}
\alias{summarise_simulation}
\title{Summarise simulation results}
\usage{
summarise_simulation(simulation_results, nsim)
}
\value{

}
\description{
Summarise simulation results
}
\author{
Shir Dekel
}
