% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_between_subjects.R
\name{unnest_between_subjects}
\alias{unnest_between_subjects}
\title{Unnest between-subjects components
Expand each between-subjects condition. Also, unnest project variation so
that each within-subjects combination (of similarity and valence) gets a
different project variation sequencej.}
\usage{
unnest_between_subjects(between_subjects_mutated)
}
\value{

}
\description{
Unnest between-subjects components
Expand each between-subjects condition. Also, unnest project variation so
that each within-subjects combination (of similarity and valence) gets a
different project variation sequencej.
}
\author{
Shir Dekel
}
