% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_project_pairs.R
\name{unnest_project_pairs}
\alias{unnest_project_pairs}
\title{Unnest project pairs
Expand remaining components so that each project type condition gets two
individual projects, and each one is associated with an anecdote variation
condition. Feature doesn't need another unnesting because we use the same
one for both target and comparison.}
\usage{
unnest_project_pairs(project_type_unnested)
}
\value{

}
\description{
Unnest project pairs
Expand remaining components so that each project type condition gets two
individual projects, and each one is associated with an anecdote variation
condition. Feature doesn't need another unnesting because we use the same
one for both target and comparison.
}
\author{
Shir Dekel
}
