% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_project_type.R
\name{unnest_project_type}
\alias{unnest_project_type}
\title{Unnest project type
Expand project type conditions with business name, type, location, and
feature so that each project type condition is associated with one of two
business names and the relevant type, location, and feature set. Integration
and structure are first unnested here and not in the previous unnesting
because they don't have five variations.}
\usage{
unnest_project_type(within_subjects_components_fixed)
}
\value{

}
\description{
Unnest project type
Expand project type conditions with business name, type, location, and
feature so that each project type condition is associated with one of two
business names and the relevant type, location, and feature set. Integration
and structure are first unnested here and not in the previous unnesting
because they don't have five variations.
}
\author{
Shir Dekel
}
