% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_project_variation.R
\name{unnest_project_variation}
\alias{unnest_project_variation}
\title{Unnest project variation
Expand sets of five (project variations)
Expand project variation with business name, type, and location, etc. to
align each set of five project variations with a business name set (of two)
and the relevant type and location.}
\usage{
unnest_project_variation(within_subjects_unnested)
}
\value{

}
\description{
Unnest project variation
Expand sets of five (project variations)
Expand project variation with business name, type, and location, etc. to
align each set of five project variations with a business name set (of two)
and the relevant type and location.
}
\author{
Shir Dekel
}
