% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_within_subjects.R
\name{unnest_within_subjects}
\alias{unnest_within_subjects}
\title{Unnest within-subjects
Expand each within-subjects condition alongside anecdote variation to show
feature type. This give each feature type condition (target or anecdote) two
anecdote variation combinations (1,2 or 2,1). Doing this before any further
business name/project variation unnesting guarantees that each combination
of conditions gets the same combination of business names. Location needs to
be unnested here so that each feature type condition gets the relevant
location (i.e., location_anecdote with anecdote).}
\usage{
unnest_within_subjects(project_variation_mutated)
}
\value{

}
\description{
Unnest within-subjects
Expand each within-subjects condition alongside anecdote variation to show
feature type. This give each feature type condition (target or anecdote) two
anecdote variation combinations (1,2 or 2,1). Doing this before any further
business name/project variation unnesting guarantees that each combination
of conditions gets the same combination of business names. Location needs to
be unnested here so that each feature type condition gets the relevant
location (i.e., location_anecdote with anecdote).
}
\author{
Shir Dekel
}
