% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unselect_for_pivot.R
\name{unselect_for_pivot}
\alias{unselect_for_pivot}
\title{Unselect for subsequent pivoting
Needs to be removed because otherwise there are NAs after pivoting}
\usage{
unselect_for_pivot(data)
}
\value{

}
\description{
Unselect for subsequent pivoting
Needs to be removed because otherwise there are NAs after pivoting
}
\author{
Shir Dekel
}
