% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anecdotes2.R
\docType{package}
\name{anecdotes2}
\alias{anecdotes2}
\alias{anecdotes2-package}
\title{Anecdotes 2 experiment}
\description{
Data, materials, and experiment code for the Anecdotes 2 experiment.
}
\details{
anecdotes2
A data package for the Anecdotes 2 experiment.

Use \code{data(package='anecdotes2')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{plot_simulation}
\link{power}
}
